requirements/base.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/

Django==1.2.3
Pinax==0.9a1
ipdb

django-debug-toolbar==0.8.3
django-staticfiles==0.2.0
django-mailer==0.2a1.dev1
django-email-confirmation==0.2.dev4
django-timezones==0.2.dev1
django-ajax-validation==0.1.3
pytz==2010h
django-openid==0.2.3.dev1
python-openid==2.2.5
django-uni-form==0.7.0
django-announcements==0.1.2
django-pagination==1.0.7
django-notification==0.2a1.dev4
idios==0.1.dev6
south
MySQL-python
BeautifulSoup==3.2.0
django-admin-tools==0.4.0
django-easy-maps==0.7
django-extensions
django-grappelli==2.2
django-mailchimp==0.1.27
django-popularity==0.2
django-taggit==0.9.1
geopy==0.94
PIL==1.1.7
python-dateutil==1.2
redis==2.2.2
sorl-thumbnail==11.01
django-endless-pagination
django-appmedia
django-classy-tags
django-cms==2.1.3
django-tinymce==1.5.1a1
#django-filebrowser==3.1
